requirements.txt
playwright>=1.42.0,<2.0.0
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
loguru>=0.7.0,<0.8.0
tenacity>=8.2.0,<9.0.0
imapclient>=2.3.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
httpx>=0.24.0,<0.27.0
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
types-pytz>=2023.3.0.0
